refactor(compat): convert TS to JS with JSDoc
Used GPT-o3-mini-high and refined manually.
